+++ examples/resources/catalystcenter_apply_credentials/resource.tf
@@ -1,5 +1,0 @@
-resource "catalystcenter_apply_credentials" "example" {
-  site_id              = "5e6f7b3a-2b0b-4a7d-8b1c-0d4b1cd5e1b1"
-  device_credential_id = "5e6f7b3a-2b0b-4a7d-8b1c-0d4b1cd5e1b1"
-  configure_device     = false
-}